git-svn-id: svn://192.168.1.10/public/CoffeeMud@10285 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Abilities/Common/CommonSkill.java
@@ -341,7 +341,7 @@ public class CommonSkill extends StdAbility
 		List<Integer> resources=room.resourceChoices();
 		if(resources!=null)
 		for(int i=0;i<resources.size();i++)
-			if(isMaterial&&(resource==(((Integer)resources.get(i)).intValue()&RawMaterial.MATERIAL_MASK)))
+			if(isMaterial&&(resource==(resources.get(i).intValue()&RawMaterial.MATERIAL_MASK)))
 				return true;
 			else
 			if(resources.get(i).equals(I))
