+++ com/planet_ink/coffee_mud/Abilities/Fighter/Fighter_ImprovedThrowing.java
@@ -68,7 +68,7 @@ public class Fighter_ImprovedThrowing extends FighterSkill
 			Item myWeapon=((MOB)affected).fetchWieldedItem();
 			if((myWeapon instanceof Weapon)
 			&&(((Weapon)myWeapon).weaponClassification()==Weapon.CLASS_THROWN))
-				affectableStats.setAttackAdjustment(affectableStats.attackAdjustment()+(int)Math.round((15.0+(float)getXLEVELLevel(invoker()))*(CMath.div(proficiency(),100.0))));
+				affectableStats.setAttackAdjustment(affectableStats.attackAdjustment()+(int)Math.round((15.0+getXLEVELLevel(invoker()))*(CMath.div(proficiency(),100.0))));
 		}
 	}
