+++ com/planet_ink/coffee_mud/Abilities/Spells/Spell_Frenzy.java
@@ -49,7 +49,7 @@ public class Spell_Frenzy extends Spell
 		if((invoker==null)&&(affected instanceof MOB))
 			invoker=(MOB)affected;
 		int xlvl=super.getXLEVELLevel(invoker());
-		float f=(float)0.1*(float)xlvl;
+		float f=(float)0.1*xlvl;
 		if(f>5.0) f=5.0f;
 		affectableStats.setDamage(affectableStats.damage()+(int)Math.round(CMath.div(affectableStats.damage(),6.0-f)));
 		affectableStats.setAttackAdjustment(affectableStats.attackAdjustment()+(int)Math.round(CMath.div(affectableStats.attackAdjustment(),6.0-f)));
