+++ com/planet_ink/coffee_mud/Abilities/Thief/Thief_SilentGold.java
@@ -60,7 +60,7 @@ public class Thief_SilentGold extends ThiefSkill
 			{
 				lastMsg=msg;
 				double money=CMLib.beanCounter().getTotalAbsoluteNativeValue(msg.source());
-				double exper=(double)getXLEVELLevel((MOB)affected);
+				double exper=getXLEVELLevel((MOB)affected);
 				double gold=money/10.0*((2.0+exper)/2);
 				if(gold>0.0)
 				{
